refactor: decompose app/page.tsx and MilkdownEditor into focused units (#747)
- Extract useSelectionTracking hook (lib/editor-page/use-selection-tracking.ts)
  from MilkdownEditor — selection count tracking is now self-contained with its
  own debounce timer and event listeners
- Extract useSaveToast hook (lib/editor-page/use-save-toast.ts) from page.tsx —
  save-complete toast visibility state and timer management is isolated
- Extract SidebarPanel component (components/SidebarPanel.tsx) from page.tsx —
  the renderPanel switch statement (files/explorer/search/outline/characters/
  dictionary/wordfreq) is now a standalone component; page.tsx passes a single
  sidebarPanelProps object rather than inlining per-panel JSX

npx tsc --noEmit passes with zero errors.
